requirements.txt
# Portfolio Flask Application - Requirements
# Install with: pip install -r requirements.txt

Flask==3.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
Werkzeug==3.0.1
